Login ve user crud işlemlerine başlandı.

+++ SekreteryaRandevu/Controllers/SirketController.cs
@@ -17,32 +17,46 @@ namespace SekreteryaRandevu.Controllers
         // GET: Sirket
         public ActionResult Liste()
         {
-            var sirkets = db.sirkets;
-            return View(sirkets.ToList());
+            if (Session["id"] != null)
+            {
+                var sirkets = db.sirkets;
+                return View(sirkets.ToList());
+            }
+            return RedirectToAction("Login", "users");
         }
 
         // GET: Sirket/Details/5
         public ActionResult Detay(int? id)
         {
-            if (id == null)
+            if (Session["id"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                sirket sirket = db.sirkets.Find(id);
+                adre adres = db.adres.Where(m => m.adresID == sirket.sirketAdresID).SingleOrDefault();
+                ViewBag.sirketAdres = adres;
+                if (sirket == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(sirket);
             }
-            sirket sirket = db.sirkets.Find(id);
-            adre adres = db.adres.Where(m => m.adresID == sirket.sirketAdresID).SingleOrDefault();
-            ViewBag.sirketAdres = adres;         
-            if (sirket == null)
-            {
-                return HttpNotFound();
-            }
-            return View(sirket);
+            return RedirectToAction("Login", "users");
+
         }
 
         // GET: Sirket/Create
         public ActionResult Olustur()
         {
-            ViewBag.ulkeler = new SelectList(db.ulkes, "ulkeID", "ulkeAdi");
-            return View();
+            if (Session["yetki"] != null && (Session["yetki"].ToString() == "3" || Session["yetki"].ToString() == "4"))
+            {
+                ViewBag.ulkeler = new SelectList(db.ulkes, "ulkeID", "ulkeAdi");
+                return View();
+            }
+            return RedirectToAction("Login", "users");
+
         }
 
         // POST: Sirket/Create
@@ -52,32 +66,42 @@ namespace SekreteryaRandevu.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Olustur(sirket sirket)
         {
-            if (ModelState.IsValid)
+            if (Session["yetki"] != null && (Session["yetki"].ToString() == "3" || Session["yetki"].ToString() == "4"))
             {
-                db.sirkets.Add(sirket);
-                db.SaveChanges();
-                return RedirectToAction("Liste");
-            }
+                if (ModelState.IsValid)
+                {
+                    db.sirkets.Add(sirket);
+                    db.SaveChanges();
+                    return RedirectToAction("Liste");
+                }
 
-            ViewBag.sirketAdresID = new SelectList(db.adres, "adresID", "adresIlce", sirket.sirketAdresID);
-            return View(sirket);
+                ViewBag.sirketAdresID = new SelectList(db.adres, "adresID", "adresIlce", sirket.sirketAdresID);
+                return View(sirket);
+            }
+            return RedirectToAction("Login", "users");
+
         }
 
         // GET: Sirket/Edit/5
         public ActionResult Duzenle(int? id)
         {
-            if (id == null)
+            if (Session["yetki"] != null && (Session["yetki"].ToString() == "3" || Session["yetki"].ToString() == "4"))
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                sirket sirket = db.sirkets.Find(id);
+                if (sirket == null)
+                {
+                    return HttpNotFound();
+                }
+                ViewBag.ulkeler = new SelectList(db.ulkes, "ulkeID", "ulkeAdi", sirket.adre.adresUlkeID);
+                ViewBag.sehirler = new SelectList(db.sehirs, "sehirID", "sehirAdi", sirket.adre.adresILID);
+                return View(sirket);
             }
-            sirket sirket = db.sirkets.Find(id);
-            if (sirket == null)
-            {
-                return HttpNotFound();
-            }
-            ViewBag.ulkeler = new SelectList(db.ulkes, "ulkeID", "ulkeAdi", sirket.adre.adresUlkeID);
-            ViewBag.sehirler = new SelectList(db.sehirs, "sehirID", "sehirAdi", sirket.adre.adresILID);
-            return View(sirket);
+            return RedirectToAction("Login", "users");
+
         }
 
         // POST: Sirket/Edit/5
@@ -86,52 +110,62 @@ namespace SekreteryaRandevu.Controllers
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Duzenle(sirket sirket)
-        {   
-            for (int i = 0; i <  sirket.iletisimToSirkets.Count(); i++)
+        {
+            if (Session["yetki"] != null && (Session["yetki"].ToString() == "3" || Session["yetki"].ToString() == "4"))
             {
-                sirket.iletisimToSirkets.ToList()[i].sirketID = sirket.sirketID;
+                for (int i = 0; i < sirket.iletisimToSirkets.Count(); i++)
+                {
+                    sirket.iletisimToSirkets.ToList()[i].sirketID = sirket.sirketID;
 
-            }            
-            for (int i = 0; i < sirket.iletisimToSirkets.Count(); i++)
-            {
-                db.Entry(sirket.iletisimToSirkets.ToList()[i]).State = EntityState.Modified;
+                }
+                for (int i = 0; i < sirket.iletisimToSirkets.Count(); i++)
+                {
+                    db.Entry(sirket.iletisimToSirkets.ToList()[i]).State = EntityState.Modified;
+                }
+                if (sirket != null)
+                {
+                    db.Entry(sirket.adre).State = EntityState.Modified;
+                    db.Entry(sirket).State = EntityState.Modified;
+                    db.SaveChanges();
+                    return RedirectToAction("Liste");
+                }
+                ViewBag.ulkeler = new SelectList(db.ulkes, "ulkeID", "ulkeAdi", sirket.adre.adresUlkeID);
+                ViewBag.sehirler = new SelectList(db.sehirs, "sehirID", "sehirAdi", sirket.adre.adresILID);
+                return View(sirket);
             }
-            if (sirket!=null)
-            {
-                db.Entry(sirket.adre).State = EntityState.Modified;
-                db.Entry(sirket).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Liste");
-            }
-            ViewBag.ulkeler = new SelectList(db.ulkes, "ulkeID", "ulkeAdi", sirket.adre.adresUlkeID);
-            ViewBag.sehirler = new SelectList(db.sehirs, "sehirID", "sehirAdi", sirket.adre.adresILID);
-            return View(sirket);
+            return RedirectToAction("Login", "users");
+
         }
 
         // GET: Sirket/Delete/5
-        
+
         public ActionResult Sil(int? id)
         {
-            if (id == null)
+            if (Session["yetki"] != null && (Session["yetki"].ToString() == "3" || Session["yetki"].ToString() == "4"))
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            sirket sirket = db.sirkets.Find(id);
-            adre adres = db.adres.Where(m => m.adresID == sirket.sirketAdresID).FirstOrDefault();
-            List<iletisimToSirket> iletisimler = db.iletisimToSirkets.Where(m => m.sirketID == id).ToList();
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                sirket sirket = db.sirkets.Find(id);
+                adre adres = db.adres.Where(m => m.adresID == sirket.sirketAdresID).FirstOrDefault();
+                List<iletisimToSirket> iletisimler = db.iletisimToSirkets.Where(m => m.sirketID == id).ToList();
 
-            if (sirket == null || adres ==null || iletisimler == null)
-            {
-                return HttpNotFound();
+                if (sirket == null || adres == null || iletisimler == null)
+                {
+                    return HttpNotFound();
+                }
+                for (int i = 0; i < iletisimler.Count(); i++)
+                {
+                    db.iletisimToSirkets.Remove(iletisimler[i]);
+                }
+                db.adres.Remove(adres);
+                db.sirkets.Remove(sirket);
+                db.SaveChanges();
+                return RedirectToAction("Liste");
             }
-            for (int i = 0; i < iletisimler.Count(); i++)
-            {
-                db.iletisimToSirkets.Remove(iletisimler[i]);
-            }
-            db.adres.Remove(adres);
-            db.sirkets.Remove(sirket);
-            db.SaveChanges();
-            return RedirectToAction("Liste");
+            return RedirectToAction("Login", "users");
+
         }
         protected override void Dispose(bool disposing)
         {
